Created a new configuration.ts file because I expect more code to be needed in this area in the future.
