+++ src/extension.ts
@@ -6,6 +6,9 @@ import {
   deactivateLanguageServerClient,
 } from "./client";
 import {
+  updateConfigurationOption
+} from "./configuration"
+import {
   initializeCodeCoverageFeatures,
   createCoverageStatusBar,
   toggleCoverageAction,
@@ -53,7 +56,6 @@ import {
   addLaunchConfiguration,
   addSettingsFileFilter,
   checkIfInstallationIsOK,
-  configFilename,
   executeClicastCommand,
   initializeInstallerFiles,
   vcastCommandtoUse,
@@ -401,9 +403,12 @@ function configureExtension(context: vscode.ExtensionContext) {
     context.subscriptions
   );
 
-  const fs = require("fs");
 
   vscode.workspace.onDidChangeConfiguration((event) => {
+    // This function gets triggered when any option at any level (user, workspace, etc.)
+    // gets changed.  The event parameter does not indicate what level has been
+    // edited but you can use the 
+
     if (event.affectsConfiguration("vectorcastTestExplorer.decorateExplorer")) {
       updateExploreDecorations();
     }
@@ -411,26 +416,7 @@ function configureExtension(context: vscode.ExtensionContext) {
       adjustVerboseSetting();
     }
     else if (event.affectsConfiguration("vectorcastTestExplorer.configurationLocation")){
-      // need to check that we have a valid path to a config file
-      const settings = vscode.workspace.getConfiguration("vectorcastTestExplorer");
-      const currentConfiguration = settings.get("configurationLocation", "");
-
-      // empty is valid, and no processing is needed
-      if (currentConfiguration.length>0) {
-        if (!fs.existsSync (currentConfiguration)) {
-          vscode.window.showErrorMessage(`Provided file path: ${currentConfiguration} does not exist`);
-          // clear illegal value 
-          settings.update ("configurationLocation", "", vscode.ConfigurationTarget.Workspace);
-        } 
-        else if (!currentConfiguration.endsWith (configFilename)) {
-          vscode.window.showErrorMessage(`Provided file path: ${currentConfiguration} is invalid (path must end with ${configFilename})`);
-          // clear illegal value
-          settings.update ("configurationLocation", "", vscode.ConfigurationTarget.Workspace);
-        }
-        else {
-          vscode.window.showInformationMessage (`Default configuration file now set to: ${currentConfiguration})`);
-        }
-      }
+      updateConfigurationOption (event);
     }
     else if (
       event.affectsConfiguration(
